requirements.txt
apache-airflow==2.10.5
boto3>=1.20.0
pandas>=1.3.0
pyarrow>=5.0.0
kafka-python>=2.0.0
botocore>=1.23.0
kafka-python==2.2.7
sqlalchemy==1.4.54
selenium==4.33.0
beautifulsoup4==4.13.4
openai==1.86.0
